update core, remove relocated classes
