+++ build.sbt
@@ -18,7 +18,7 @@ val logbackVersion             = "1.5.11"
 val log4catsVersion            = "2.7.0"
 val lucumaCatalogVersion       = "0.48.5"
 val lucumaItcVersion           = "0.22.5"
-val lucumaCoreVersion          = "0.106.1"
+val lucumaCoreVersion          = "0.106.2"
 val lucumaGraphQLRoutesVersion = "0.8.15"
 val lucumaSsoVersion           = "0.6.26"
 val munitVersion               = "0.7.29" // check test output if you attempt to update this
