+++ modules/service/src/main/scala/lucuma/odb/service/ObservationWorkflowService.scala
@@ -15,6 +15,7 @@ import lucuma.core.enums.Instrument
 import lucuma.core.enums.ObsActiveStatus
 import lucuma.core.enums.ObservationExecutionState
 import lucuma.core.enums.ObservationValidationCode
+import lucuma.core.enums.ObservationWorkflowState
 import lucuma.core.enums.ScienceBand
 import lucuma.core.enums.Site
 import lucuma.core.math.Coordinates
@@ -25,6 +26,7 @@ import lucuma.core.model.ConfigurationRequest
 import lucuma.core.model.ObjectTracking
 import lucuma.core.model.Observation
 import lucuma.core.model.ObservationValidation
+import lucuma.core.model.ObservationWorkflow
 import lucuma.core.model.ObservingNight
 import lucuma.core.model.Program
 import lucuma.core.model.StandardRole.*
@@ -33,8 +35,6 @@ import lucuma.core.util.DateInterval
 import lucuma.core.util.Enumerated
 import lucuma.itc.client.ItcClient
 import lucuma.odb.data.ObservationValidationMap
-import lucuma.odb.data.ObservationWorkflow
-import lucuma.odb.data.ObservationWorkflowState
 import lucuma.odb.data.Tag
 import lucuma.odb.graphql.enums.Enums
 import lucuma.odb.logic.TimeEstimateCalculatorImplementation
